Project update/rework
A nearly complete rework of the project, notable changes are a complete rework of the points-system and removal of the camera-follow-bullet-system as well as major parts of the bullet-time system. Additions are more varied procedural generation in terms of pillar placement, as well as some programmatic VFX (camera shake on pillar hit, pillars getting tilted when hit), and the ability to rotate the player around the scene for differing shot-opportunities.